+++ fractal.config.js
@@ -14,13 +14,13 @@ const pkg = require(path.join(__dirname, 'package.json'));
 // Set the title
 fractal.set('project.title', 'Mozilla Protocol');
 
-// Provide the package.json "version" to the templates
+// Provide the package.json 'version' to the templates
 fractal.set('project.version', pkg.version);
 
 // Server configuration
-fractal.web.set("server.port", 3000);
-fractal.web.set("server.sync", true);
-fractal.web.set("debug", true);
+fractal.web.set('server.port', 3000);
+fractal.web.set('server.sync', true);
+fractal.web.set('debug', true);
 
 // Configure components
 fractal.components.engine(nunjucks);
@@ -29,14 +29,14 @@ fractal.components.set('path', path.join(__dirname, 'components'));
 fractal.components.set('default.status', null);
 fractal.components.set('statuses', {
     deprecated: {
-        label: "Deprecated",
-        description: "Deprecated component slated for removal.",
-        color: "firebrick"
+        label: 'Deprecated',
+        description: 'Deprecated component slated for removal.',
+        color: 'firebrick'
     },
     wip: {
-        label: "WIP",
-        description: "Work in progress, not yet ready for use.",
-        color: "darkorange"
+        label: 'WIP',
+        description: 'Work in progress, not yet ready for use.',
+        color: 'darkorange'
     }
 });
 
@@ -45,14 +45,14 @@ fractal.docs.set('ext', '.md');
 fractal.docs.set('path', path.join(__dirname, 'docs'));
 fractal.docs.set('statuses', {
     outdated: {
-        label: "Outdated",
-        description: "Outdated documentation, needs updating.",
-        color: "firebrick"
+        label: 'Outdated',
+        description: 'Outdated documentation, needs updating.',
+        color: 'firebrick'
     },
     draft: {
-        label: "Draft",
-        description: "Draft documentation, work in progress.",
-        color: "darkorange"
+        label: 'Draft',
+        description: 'Draft documentation, work in progress.',
+        color: 'darkorange'
     }
 });
 
@@ -60,7 +60,7 @@ fractal.docs.set('statuses', {
 fractal.web.set('static.path', path.join(__dirname, 'static'));
 
 // Tell Fractal where to build the site.
-fractal.web.set('builder.dest', __dirname + 'dist');
+fractal.web.set('builder.dest', path.join(__dirname, 'dist'));
 
 // Custom theme
 const theme = require('./theme');